Refactoring "Models" to their intended scope
Apologies for the size of this PR. Can't be helped. I know it's probably
going to be hard to read, LMK if you have questions.
- Moved objects out of the "Models" namespace. Almost everything in Core
is a "model", so the differentiation is unnecessary.
- Moved Management API classes that were *only* for managing API inputs
to the Management API project.
- Broke up a couple files that violated the "1 class per file" rule.
- Moved base classes and Management classes into separate folders, but
not separate namespaces (at least for now). To help people looking at
the source better understand the taxonomy.
- Moved any classes that didn't seem to fill either one of those buckets
into the root of Core. Core objects like User should stay there. But we
won't know fully which files are Management-only until the AuthAPI is
cooked.
- Moved FileUploadParameter to Core.ApiClients, so it can be used by
both Client projects.
- Added "InternalsVisibleTo" attribute to Core/Assembly.cs
- Fixed namespace references due to refactor, and "Remove & Sort"-ed
many Usings sections.